Update course.js
WARNING! I'm having troubling connecting to the server right now so this code is not tested.  I am so sorry, I'll test it as soon as I can connect to the database or if someone else can test and knows how to fix this potentially broken code/ confirm it's OK before I can connect that's also good!